+++ js/test/date/testDayOfWeek.js
@@ -2049,6 +2049,16 @@ module.exports.testweekdata = {
         test.equal(info.getWeekEndEnd(), 0);
         test.done();
     },
+    testWeekData_my_MM: function(test) {
+        test.expect(4);
+        var info = new LocaleInfo("my-MM");
+        test.ok(info !== null);
+
+        test.equal(info.getFirstDayOfWeek(), 0);
+        test.equal(info.getWeekEndStart(), 6);
+        test.equal(info.getWeekEndEnd(), 0);
+        test.done();
+    },
     testWeekData_ne_NP: function(test) {
         test.expect(4);
         var info = new LocaleInfo("ne-NP");