+++ js/test/root/testscriptinfo.js
@@ -2266,6 +2266,16 @@ module.exports.testscriptinfo = {
         test.equal(scinfo.getScriptDirection(), "ltr");
         test.done();
     },
+    testScriptInfo_my_MM: function(test) {
+        test.expect(4);
+        var li = new LocaleInfo("my-MM");
+        var scinfo = new ScriptInfo(li.getScript());
+        test.ok(li !== null);
+        test.ok(scinfo !== null);
+        test.equal(li.getScript(), "Mymr");
+        test.equal(scinfo.getScriptDirection(), "ltr");
+        test.done();
+    },
     testScriptInfo_ne_NP: function(test) {
         test.expect(4);
         var li = new LocaleInfo("ne-NP");