+++ js/test/root/teststrings.js
@@ -3352,6 +3352,27 @@ module.exports.teststrings = {
         test.equal(str.formatChoice(16), "Default items");
         test.done();
     },
+    testStringFormatChoiceCharClasses_my_MM: function(test) {
+        test.expect(2);
+        var str = new IString("0#There are no items.|one#There items are one|few#The items are few|many#The items are many|#Default items");
+        str.setLocale("my-MM");
+
+        test.ok(str !== null);
+
+        test.equal(str.formatChoice(1), "There items are one");
+        test.done();
+    },
+    testStringFormatChoiceCharClasses_my_MM2: function(test) {
+        test.expect(2);
+        var str = new IString("0#There are no items.|one#There items are one|few#The items are few|many#The items are many|#Default items");
+        str.setLocale("my-MM");
+
+        test.ok(str !== null);
+
+        test.equal(str.formatChoice(5), "Default items");
+        test.done();
+    },
+    
     testStringFormatChoiceCharClassesComplex1: function(test) {
         test.expect(2);
         var str = new IString("0#There are no items.|one#There items are one|few#The items are few|many#The items are many|#Default items");
@@ -3949,6 +3970,26 @@ module.exports.teststrings = {
         test.ok(str !== null);
 
         test.equal(str.formatChoice(1.7), "Default items");
+        test.done();
+    },
+    testStringFormatChoiceDecimal_my_MM: function(test) {
+        test.expect(2);
+        var str = new IString("0#There are no items.|one#There items are one|few#The items are few|many#The items are many|#Default items");
+        str.setLocale("my-MM");
+
+        test.ok(str !== null);
+
+        test.equal(str.formatChoice(1.0), "There items are one");
+        test.done();
+    },
+    testStringFormatChoiceDecimal_my_MM2: function(test) {
+        test.expect(2);
+        var str = new IString("0#There are no items.|one#There items are one|few#The items are few|many#The items are many|#Default items");
+        str.setLocale("my-MM");
+
+        test.ok(str !== null);
+
+        test.equal(str.formatChoice(3.7), "Default items");
         test.done();
     },
     testStringFormatChoiceDecimal_ne_NP: function(test) {