+++ js/test/units/testSuiteFiles.js
@@ -40,6 +40,7 @@ module.exports.files = [
     "testunitfmt_ky_KG.js",
     "testunitfmt_lo_LA.js",
     "testunitfmt_ne_NP.js",
+    "testunitfmt_my_MM.js",
     "testunitfmtasync.js",
     "testdigitalspeed.js",
     "testunitfmt_usages.js"