+++ docs/ReleaseNotes.md
@@ -11,7 +11,8 @@ Bug Fixes:
 * Updated to IANA time zone data 2020a
 * Put base option back on ResBundle
 * Fixed a few tools and dependencies so that ilib works on node v14
-* Fixed a bug which a default script for ky should be Cyrl instead of Arabic
+* Fixed a default timezone in `my-MM` to `Asia/Yangon`
+* Fixed a bug which a default script for `ky` should be `Cyrl` instead of `Arabic`
 
 Build 009
 -------